+++ dom/media/webrtc/RTCCertificate.cpp
@@ -11,6 +11,7 @@
 #include "jsapi.h"
 #include "mozilla/dom/CryptoKey.h"
 #include "mozilla/dom/RTCCertificateBinding.h"
+#include "mozilla/dom/StructuredCloneHolder.h"
 #include "mozilla/dom/WebCryptoCommon.h"
 #include "mozilla/dom/WebCryptoTask.h"
 #include "mozilla/Move.h"
@@ -311,7 +312,7 @@ bool RTCCertificate::WritePrivateKey(JSStructuredCloneWriter* aWriter) const {
   if (!jwk.ToJSON(json)) {
     return false;
   }
-  return WriteString(aWriter, json);
+  return StructuredCloneHolder::WriteString(aWriter, json);
 }
 
 bool RTCCertificate::WriteCertificate(JSStructuredCloneWriter* aWriter) const {
@@ -326,7 +327,7 @@ bool RTCCertificate::WriteCertificate(JSStructuredCloneWriter* aWriter) const {
 }
 
 bool RTCCertificate::WriteStructuredClone(
-    JSStructuredCloneWriter* aWriter) const {
+    JSContext* aCx, JSStructuredCloneWriter* aWriter) const {
   if (!mPrivateKey || !mCertificate) {
     return false;
   }
@@ -339,7 +340,7 @@ bool RTCCertificate::WriteStructuredClone(
 
 bool RTCCertificate::ReadPrivateKey(JSStructuredCloneReader* aReader) {
   nsString json;
-  if (!ReadString(aReader, json)) {
+  if (!StructuredCloneHolder::ReadString(aReader, json)) {
     return false;
   }
   JsonWebKey jwk;
@@ -363,21 +364,33 @@ bool RTCCertificate::ReadCertificate(JSStructuredCloneReader* aReader) {
   return !!mCertificate;
 }
 
-bool RTCCertificate::ReadStructuredClone(JSStructuredCloneReader* aReader) {
+// static
+already_AddRefed<RTCCertificate> RTCCertificate::ReadStructuredClone(
+    JSContext* aCx, nsIGlobalObject* aGlobal,
+    JSStructuredCloneReader* aReader) {
+  if (!NS_IsMainThread()) {
+    // These objects are mainthread-only.
+    return nullptr;
+  }
   uint32_t version, authType;
   if (!JS_ReadUint32Pair(aReader, &version, &authType) ||
       version != RTCCERTIFICATE_SC_VERSION) {
-    return false;
+    return nullptr;
   }
-  mAuthType = static_cast<SSLKEAType>(authType);
+  RefPtr<RTCCertificate> cert = new RTCCertificate(aGlobal);
+  cert->mAuthType = static_cast<SSLKEAType>(authType);
 
   uint32_t high, low;
   if (!JS_ReadUint32Pair(aReader, &high, &low)) {
-    return false;
+    return nullptr;
   }
-  mExpires = static_cast<PRTime>(high) << 32 | low;
+  cert->mExpires = static_cast<PRTime>(high) << 32 | low;
 
-  return ReadPrivateKey(aReader) && ReadCertificate(aReader);
+  if (!cert->ReadPrivateKey(aReader) || !cert->ReadCertificate(aReader)) {
+    return nullptr;
+  }
+
+  return cert.forget();
 }
 
 }  // namespace dom