+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -8,6 +8,13 @@ exclude: true
 
 <h1> GeckoView API Changelog. </h1>
 
+## v69
+- Modified behavior of ['setAutomaticFontSizeAdjustment'][69.1] so that it no 
+  longer has any effect on ['setFontInflationEnabled'][69.2]
+
+[69.1]: ./GeckoRuntimeSettings.html#setAutomaticFontSizeAdjustment-boolean-
+[69.2]: ./GeckoRuntimeSettings.html#setFontInflationEnabled-boolean-
+
 ## v68
 - Added [`GeckoRuntime#configurationChanged`][68.1] to notify the device
   configuration has changed.